refactor code to add float

+++ Calculator.rb
@@ -1,6 +1,6 @@
 puts "Enter a number: "
-num1 = gets.chomp.to_i
+num1 = gets.chomp.to_f
 puts "Enter another number: "
-num2 = gets.chomp.to_i
+num2 = gets.chomp.to_f
 
 puts "Adding these numbers will give us #{num1 + num2}"
